backend/requirements.txt
tensorflow==2.10.0
keras==2.10.0
numpy==1.23.5
matplotlib==3.6.2
pillow==9.3.0
scikit-learn==1.1.3
pandas==1.5.1
h5py==3.7.0

# API requirements
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.5
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
passlib==1.7.4
typing-extensions==4.7.1